+++ Nickel/Framework/Implementations/AssemblyModLoadContextProvider.cs
@@ -1,5 +1,6 @@
 using Nanoray.PluginManager;
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
@@ -8,25 +9,15 @@ using System.Runtime.Loader;
 
 namespace Nickel;
 
-internal sealed class AssemblyModLoadContextProvider(
-	AssemblyLoadContext sharedContext
-) : IAssemblyPluginLoaderLoadContextProvider<IAssemblyModManifest>
+internal sealed class AssemblyModLoadContextProvider : IAssemblyPluginLoaderLoadContextProvider<IAssemblyModManifest>
 {
-	private AssemblyLoadContext SharedContext { get; } = sharedContext;
-	private ConditionalWeakTable<IPluginPackage<IAssemblyModManifest>, WeakReference<AssemblyLoadContext>> ContextCache { get; } = [];
+	private readonly AssemblyLoadContext? FallbackContext;
+	private readonly ConditionalWeakTable<IPluginPackage<IAssemblyModManifest>, WeakReference<AssemblyLoadContext>> ContextCache = [];
+	private readonly List<WeakReference<Assembly>> SharedAssemblies = [];
 
-	private static bool ShouldShare(IPluginPackage<IAssemblyModManifest> package, AssemblyName assemblyName)
+	public AssemblyModLoadContextProvider(AssemblyLoadContext? fallbackContext)
 	{
-		var reference = package.Manifest.AssemblyReferences.FirstOrDefault(r => r.Name == (assemblyName.Name ?? assemblyName.FullName));
-		if (reference is not null)
-			return reference.IsShared;
-		if (typeof(Nickel).Assembly.GetReferencedAssemblies().Any(a => (a.Name ?? a.FullName) == (assemblyName.Name ?? assemblyName.FullName)))
-			return true;
-		if (package.PackageRoot.GetRelativeFile($"{assemblyName.Name ?? assemblyName.FullName}.dll").Exists)
-			return false;
-		if (File.Exists(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, $"{assemblyName.Name ?? assemblyName.FullName}.dll")))
-			return true;
-		return false;
+		this.FallbackContext = fallbackContext;
 	}
 
 	public AssemblyLoadContext GetLoadContext(IPluginPackage<IAssemblyModManifest> package)
@@ -34,22 +25,25 @@ internal sealed class AssemblyModLoadContextProvider(
 		if (this.ContextCache.TryGetValue(package, out var weakContext) && weakContext.TryGetTarget(out var context))
 			return context;
 
-		context = new CustomContext(this.SharedContext, package);
+		context = new CustomContext(this.FallbackContext, this.SharedAssemblies, package);
 		this.ContextCache.AddOrUpdate(package, new WeakReference<AssemblyLoadContext>(context));
 		return context;
 	}
 
 	private sealed class CustomContext : AssemblyLoadContext
 	{
-		private AssemblyLoadContext SharedContext { get; }
+		private readonly AssemblyLoadContext? FallbackContext;
+		private readonly List<WeakReference<Assembly>> SharedAssemblies;
 		private IPluginPackage<IAssemblyModManifest> Package { get; }
 
 		public CustomContext(
-			AssemblyLoadContext sharedContext,
+			AssemblyLoadContext? fallbackContext,
+			List<WeakReference<Assembly>> sharedAssemblies,
 			IPluginPackage<IAssemblyModManifest> package
 		) : base(package.Manifest.UniqueName)
 		{
-			this.SharedContext = sharedContext;
+			this.FallbackContext = fallbackContext;
+			this.SharedAssemblies = sharedAssemblies;
 			this.Package = package;
 		}
 
@@ -57,33 +51,41 @@ internal sealed class AssemblyModLoadContextProvider(
 		{
 			if (this.Assemblies.FirstOrDefault(a => (a.GetName().Name ?? a.GetName().FullName) == (assemblyName.Name ?? assemblyName.FullName)) is { } existingPrivateAssembly)
 				return existingPrivateAssembly;
-			if (this.SharedContext.Assemblies.FirstOrDefault(a => (a.GetName().Name ?? a.GetName().FullName) == (assemblyName.Name ?? assemblyName.FullName)) is { } existingSharedAssembly)
-				return existingSharedAssembly;
 
-			if (ShouldShare(this.Package, assemblyName))
+			if (this.FallbackContext is { } fallbackContext)
+				if (fallbackContext.Assemblies.FirstOrDefault(a => (a.GetName().Name ?? a.GetName().FullName) == (assemblyName.Name ?? assemblyName.FullName)) is { } existingFallbackAssembly)
+					return existingFallbackAssembly;
+
+			foreach (var weakSharedAssembly in this.SharedAssemblies)
+				if (weakSharedAssembly.TryGetTarget(out var sharedAssembly) && (sharedAssembly.GetName().Name ?? sharedAssembly.GetName().FullName) == (assemblyName.Name ?? assemblyName.FullName))
+					return sharedAssembly;
+
+			var file = this.Package.PackageRoot.GetRelativeFile($"{assemblyName.Name ?? assemblyName.FullName}.dll");
+			if (file.Exists)
 			{
-				try
-				{
-					return this.SharedContext.LoadFromAssemblyName(assemblyName);
-				}
-				catch
-				{
-					using var assemblyStream = this.Package.PackageRoot.GetRelativeFile($"{assemblyName.Name ?? assemblyName.FullName}.dll").OpenRead();
-					return this.SharedContext.LoadFromStream(assemblyStream);
-				}
+				using var assemblyStream = this.Package.PackageRoot.GetRelativeFile($"{assemblyName.Name ?? assemblyName.FullName}.dll").OpenRead();
+				var assembly = this.LoadFromStream(assemblyStream);
+
+				if (this.ShouldShare(assemblyName))
+					this.SharedAssemblies.Add(new(assembly));
+
+				return assembly;
 			}
-			else
-			{
-				try
-				{
-					using var assemblyStream = this.Package.PackageRoot.GetRelativeFile($"{assemblyName.Name ?? assemblyName.FullName}.dll").OpenRead();
-					return this.LoadFromStream(assemblyStream);
-				}
-				catch
-				{
-					return this.SharedContext.LoadFromAssemblyName(assemblyName);
-				}
-			}
+
+			return this.FallbackContext?.LoadFromAssemblyName(assemblyName);
+		}
+
+		private bool ShouldShare(AssemblyName assemblyName)
+		{
+			if (this.Package.Manifest.AssemblyReferences.FirstOrDefault(r => r.Name == (assemblyName.Name ?? assemblyName.FullName)) is { } reference)
+				return reference.IsShared;
+			if (typeof(Nickel).Assembly.GetReferencedAssemblies().Any(a => (a.Name ?? a.FullName) == (assemblyName.Name ?? assemblyName.FullName)))
+				return true;
+			if (this.Package.PackageRoot.GetRelativeFile($"{assemblyName.Name ?? assemblyName.FullName}.dll").Exists)
+				return this.Package.Manifest.EntryPointAssembly == $"{assemblyName.Name ?? assemblyName.FullName}.dll";
+			if (File.Exists(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, $"{assemblyName.Name ?? assemblyName.FullName}.dll")))
+				return true;
+			return false;
 		}
 	}
 }